Error if nvidia-current could not install

+++ scripts/graphicTest/chroot/installNvidia.bash
@@ -1,5 +1,14 @@
 #!/bin/bash
 apt-get install -y nvidia-current
+if [ $? != 0 ]; then
+    echo ''
+    echo '---------------------------------------------------------'
+    echo 'Could not install the latest nvidia-driver nvidia-current'
+    echo '---------------------------------------------------------'
+    echo ''
+    exit 1
+fi
+
 rm -rf tmpNvidiaFiles; mkdir tmpNvidiaFiles; cd tmpNvidiaFiles
 arch=`dpkg --print-architecture`
 file=`ls /tmp/nvidia | grep $arch`
